+++ crates/jsonschema/src/keywords/minmax.rs
@@ -4,7 +4,6 @@ use crate::{
     ext::numeric,
     keywords::CompilationResult,
     paths::{LazyLocation, Location},
-    thread::ThreadBound,
     types::JsonType,
     validator::{Validate, ValidationContext},
 };
@@ -29,7 +28,7 @@ macro_rules! define_numeric_keywords {
 
             impl<T> Validate for $struct_name<T>
             where
-                T: Copy + ThreadBound + num_traits::ToPrimitive,
+                T: Copy + Send + Sync + num_traits::ToPrimitive,
                 u64: NumCmp<T>,
                 i64: NumCmp<T>,
                 f64: NumCmp<T>,