+++ crates/jsonschema/src/options.rs
@@ -8,7 +8,6 @@ use crate::{
     keywords::{custom::KeywordFactory, format::Format},
     paths::Location,
     retriever::DefaultRetriever,
-    thread::ThreadBound,
     Keyword, ValidationError, Validator,
 };
 use ahash::AHashMap;
@@ -396,7 +395,7 @@ impl<R> ValidationOptions<R> {
     pub fn with_format<N, F>(mut self, name: N, format: F) -> Self
     where
         N: Into<String>,
-        F: Fn(&str) -> bool + ThreadBound + 'static,
+        F: Fn(&str) -> bool + Send + Sync + 'static,
     {
         self.formats.insert(name.into(), Arc::new(format));
         self
@@ -504,7 +503,8 @@ impl<R> ValidationOptions<R> {
                 &'a Value,
                 Location,
             ) -> Result<Box<dyn Keyword>, ValidationError<'a>>
-            + ThreadBound
+            + Send
+            + Sync
             + 'static,
     {
         self.keywords.insert(name.into(), Arc::new(factory));